requirements.txt
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
beautifulsoup4==4.12.3
lxml==5.3.0
dateparser==1.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
regex==2024.9.11
rich==13.8.1